Support building dependency packages
Dependency packages are packages that are handled only for Debian package
building purposes. Their jobs are stored in a separate folder that doesn't
clutter the main swh package views.

To do so, we duplicate the gbp-buildpackage and automatic-backport jobs so they
can be created in a separate folder. This is just a matter of playing with yaml
anchors. We also adapt the incoming-tag job to look for jobs in both places.