Apply arc patches more forcefully
When arc applies a diff that's stacked on top of another one, it applies the
full stack. However, it also edits the commit message, which changes the
identifier of the base commit, which makes it fail to apply higher diffs.

Turning on the force flag allows stacks to apply properly.

Close T1445.