jenkins-job: Declare explicitely the python_module for each swh project
This allows to deal with naming discrepancies in debian/python/archive naming
as the swh-icinga-plugins demonstrates.

Its current build fails because of naming discrepancy, the:
- repository is named swh-icinga-plugins
- debian package is named python3-swh.icingaplugin
- python module is named swh.icinga-plugins
- archive pushed for the python module is name swh.icinga_plugins

So this defines an explicit extra entry name `python_module` used to start the
debian package build (out of the pypi module artifact holding that name).

This fixes the build [1]

[1] https://jenkins.softwareheritage.org/job/debian/job/packages/job/DICP/job/update-for-release/3