Simplify docker image build to not override the docker image to build
That's up to the repository's Makefile to determine that. That way, it currently fixes
the build and push tag logic target to actually work. But bonus, it will allow to keep
the building logic of image(s) within the scope of the repository which knows what to
do (so building multiple images becomes possible with that very build).

So now the makefile must respect the following conventions (still explained in the
template):
- build: locally build the image out of the repository
TAG, IMAGE_LABEL (to ease cleanup part) must be provided
- test: test the locally built image
TAG must be provided
- clean: clean up locally built image (parametric on the jenkins build)
IMAGE_LABEL must be provided
- push: publish to the registry
TAG, REGISTRY_URL must be provided